events, ip tools, http lib

+++ src/main/java/io/github/alerithe/client/ui/click/FreeFormUI.java
@@ -157,7 +157,7 @@ public class FreeFormUI extends GuiScreen {
                     intProp.getValue(), intProp.getMinimum(), intProp.getMaximum()) {
                 @Override
                 public void onValueChanged(double oldValue, double newValue) {
-                    intProp.setValue(MathHelper.floor(newValue));
+                    intProp.setValue((int) MathHelper.floor(newValue));
                 }
             };
             valueSlider.addChildren(nameLbl, valueLbl);
